Refactor: 컴포넌트 분리, 모달창 버튼 수정
