+++ package.json
@@ -10,12 +10,16 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@date-io/date-fns": "^3.2.0",
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
     "@mui/icons-material": "^6.3.1",
     "@mui/joy": "5.0.0-beta.51",
     "@mui/material": "^6.3.1",
     "@mui/styled-engine-sc": "^6.3.1",
+    "@mui/x-date-pickers": "^7.23.6",
+    "date-fns": "^4.1.0",
+    "dayjs": "^1.11.13",
     "moment": "^2.30.1",
     "react": "^18.3.1",
     "react-big-calendar": "^1.17.1",
